requirements.txt
flet>=0.21.0
pdf2image>=1.16.0
pytesseract>=0.3.10
python-pptx>=0.6.21
Pillow>=10.0.0
psutil>=5.9.0
pyinstaller>=6.0.0
